backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Session==0.5.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-api-python-client==2.95.0
python-dotenv==1.0.0
moviepy==1.0.3